New note page created
