+++ src/app/new-note/components/MenuBar.tsx
@@ -1,49 +1,10 @@
-'use client'
-import '../styles/editor.css'
-import { useState, useEffect } from 'react'
-
-import { EditorContent, BubbleMenu, Editor, useEditor } from '@tiptap/react'
-import Placeholder from '@tiptap/extension-placeholder'
-import StarterKit from '@tiptap/starter-kit'
-
 import { AiOutlineBold, AiOutlineClear, AiOutlineCode, AiOutlineItalic, AiOutlineRedo, AiOutlineStrikethrough, AiOutlineUndo } from 'react-icons/ai'
 import { TbClearFormatting, TbBlockquote } from 'react-icons/tb'
 import { BsCodeSlash, BsFileBreak, BsParagraph } from 'react-icons/bs'
 import { LuHeading1, LuHeading2, LuHeading3, LuHeading4, LuHeading5, LuHeading6 } from 'react-icons/lu'
 import { MdFormatListBulleted, MdHorizontalRule } from 'react-icons/md'
 import { FaListOl } from 'react-icons/fa'
-
-const NewNote = () => {
-    const [title, setTitle] = useState('');
-    const [isEditable, setIsEditable] = useState(true)
-    const editor = useEditor({
-        extensions: [
-            StarterKit,
-            Placeholder.configure({
-                placeholder: 'Write something …',
-            }),
-        ],
-    })
-    useEffect(() => {
-        //* If editable is true then show bubble menu
-        if (editor) {
-            editor.setEditable(isEditable)
-        }
-    }, [isEditable, editor])
-    return (
-        <main className='max-w-full'>
-            {/* <div>
-                <input type="checkbox" checked={isEditable} onChange={() => setIsEditable(!isEditable)} />
-                Editable
-            </div> */}
-
-            <MenuBar editor={editor} />
-            <input type='text' value={title} onChange={(e) => setTitle(e.target.value)} className='w-full mb-2 text-2xl font-bold outline-none sm:text-3xl md:text-4xl bg-inherit' placeholder="What's The Title" />
-            <hr className='' />
-            <TextEditor editor={editor} />
-        </main>
-    )
-}
+import { Editor } from '@tiptap/react'
 
 
 const MenuBar = ({ editor }: { editor: Editor | null }) => {
@@ -243,74 +204,4 @@ const MenuBar = ({ editor }: { editor: Editor | null }) => {
     )
 }
 
-function TextEditor({ editor }: { editor: Editor | null }) {
-    // const [isEditable, setIsEditable] = useState(true)
-    // const [content,setContent] = useState<String| null>('');
-    // const editor = useEditor({
-    //     extensions: [StarterKit],
-    //     content: ``,
-    // })
-    // useEffect(() => {
-    //     //* If editable is true then show bubble menu
-    //     if (editor) {
-    //         editor.setEditable(isEditable)
-    //     }
-    // }, [isEditable, editor])
-
-    return (
-        <div className='max-w-full prose lg:prose-lg'>
-            {/* <div>
-                <input type="checkbox" checked={isEditable} onChange={() => setIsEditable(!isEditable)} />
-                Editable
-            </div> */}
-            {editor && <BubbleMenu editor={editor} tippyOptions={{ duration: 100 }}>
-                <button
-                    onClick={() => editor.chain().focus().toggleBold().run()}
-                    className={`bg-white px-2 py-1 mx-1 drop-shadow-md outline outline-1 rounded-md ${editor.isActive('bold') && 'bg-gray-500'} ${!editor.can()
-                        .chain()
-                        .focus()
-                        .toggleItalic()
-                        .run() && 'opacity-70'}`}
-                >
-                    <AiOutlineBold className={`w-6 h-6 text-gray-500 ${editor.isActive('bold') && 'text-white'} `}/>
-                </button>
-                <button
-                    onClick={() => editor.chain().focus().toggleItalic().run()}
-                    disabled={
-                        !editor.can()
-                            .chain()
-                            .focus()
-                            .toggleItalic()
-                            .run()
-                    }
-                    className={`bg-white px-2 py-1 mx-1 drop-shadow-md outline outline-1 rounded-md ${editor.isActive('italic') && 'bg-gray-500'} ${!editor.can()
-                        .chain()
-                        .focus()
-                        .toggleItalic()
-                        .run() && 'opacity-70'}`}
-                >
-                    <AiOutlineItalic className={`w-6 h-6 text-gray-500 ${editor.isActive('italic') && 'text-white'}`} />
-                </button>
-                <button
-                    onClick={() => editor.chain().focus().toggleStrike().run()}
-                    disabled={
-                        !editor.can()
-                            .chain()
-                            .focus()
-                            .toggleStrike()
-                            .run()
-                    }
-                    className={`bg-white px-2 py-1 mx-1 drop-shadow-md outline outline-1 rounded-md ${editor.isActive('strike') && 'bg-gray-500'} ${!editor.can()
-                        .chain()
-                        .focus()
-                        .toggleItalic()
-                        .run() && 'opacity-70'}`}
-                >
-                    <AiOutlineStrikethrough className={`w-6 h-6 text-gray-500 ${editor.isActive('strike') && 'text-white'}`} />
-                </button>
-            </BubbleMenu>}
-            <EditorContent editor={editor} />
-        </div>
-    )
-}
-export default NewNote
+export default MenuBar
